+++ pe-bear/gui/pe_models/ImportsTreeModel.h
@@ -92,5 +92,5 @@ protected:
 		ORDINAL,
 		ADDED_COLS_NUM
 	};
-	uint32_t libraryId;
+	size_t libraryId;
 };
